refactor: RouteEvent should be handled by the routecontroller
