+++ src/app/renderer/pages/index.html
@@ -311,7 +311,11 @@
               class="h-100 d-flex flex-row flex-nowrap text-nowrap"
               style="overflow-x: auto"
               id="route-container"
-            ></div>
+            >
+            <div class="d-flex align-items-center justify-content-center text-center w-100 h-100 text-white" id="no-routing-data-text">
+              No Routing Data
+            </div>
+          </div>
           </div>
           <!--Settings content-->
           <div id="settings-tab-content" class="w-100 h-100 hide">
